web_channel/requirements.txt
amqp==5.2.0
asgiref==3.3.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==3.0.4
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
crispy-tailwind==0.2.0
cryptography==42.0.8
daphne==3.0.2
Django==3.1.4
django-celery-results==2.4.0
django-crispy-forms==1.10.0
django-environ==0.4.5
django-extensions==3.1.5
django-jsonfield==1.4.1
django-tailwind==1.1.0
djangorestframework==3.15.1
dnspython==2.6.1
et-xmlfile==1.1.0
eventlet==0.36.1
greenlet==3.0.3
gunicorn==20.0.4
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
kombu==5.3.7
msgpack==1.0.8
numpy==1.26.4
opencv-python==4.10.0.82
openpyxl==3.1.4
outcome==1.3.0.post0
packaging==24.1
pdfminer.six==20231228
pdfplumber==0.11.1
pika==1.3.2
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
pypdfium2==4.30.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2020.4
#pywin32==306
redis==5.0.6
requests==2.32.3
selenium==4.21.0
service-identity==24.1.0
setuptools==70.0.0
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.1
trio==0.25.1
trio-websocket==0.11.1
Twisted==24.3.0
#twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==5.2.0
wsproto==1.2.0
zope.interface==6.4.post2
